python/requirements.txt
PyQt5
qdarkstyle
pybluez
pyyaml